backend/requirements.txt
Django==6.0.2
djangorestframework==3.14.0
python-decouple==3.8
django-cors-headers==4.3.1
psycopg2-binary>=2.9.11
Pillow>=11.0.0
